backend/pyproject.toml
[project]
name = "llm-omnius"
version = "0.1.0"
description = "Automated news framing analysis tool based on Robert Entman's methodology using Groq LLM intelligence."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "langdetect",
    "groq",
    "pydantic",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "tenacity",
    "pydantic-ai",
    "tavily-python",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]
